refactor: 추상화 명명 수정
