+++ src/main/java/save/DatabaseConnection.java
@@ -1,0 +1,8 @@
+package save;
+
+import java.sql.Connection;
+
+public interface DatabaseConnection {
+
+    Connection getConnection();
+}
