+++ src/test/java/save/TestMySQLConnection.java
@@ -5,7 +5,7 @@ import java.sql.DriverManager;
 import java.sql.SQLException;
 import java.sql.Statement;
 
-public class TestJanggiConnection implements MySQLConnection {
+public class TestMySQLConnection implements DatabaseConnection {
 
     private static final String SERVER = "localhost:13306"; // MySQL 서버 주소
     private static final String DATABASE = "test_janggi"; // MySQL DATABASE 이름
@@ -13,7 +13,7 @@ public class TestJanggiConnection implements MySQLConnection {
     private static final String USERNAME = "root"; //  MySQL 서버 아이디
     private static final String PASSWORD = "root"; // MySQL 서버 비밀번호
 
-    public TestJanggiConnection() {
+    public TestMySQLConnection() {
         createDatabaseIfNotExists();
     }
